Refactors ScraperPlugin to Plugin
Renames ScraperPlugin to Plugin for better generalization.

Adds a HeaderButton method to the Plugin class to allow
plugins to define a button that appears in the header of the
plugin's view.

Updates plugins to use the new Plugin class and adds
implementation for the HeaderButton method to trigger the
scraping process.

Changes `scrap` method to `getHtmlDocument` in InternetRequest class
to better reflect its purpose.